Port to current version of escpos-printer-db.
